+++ shell/browser/browser_win.cc
@@ -369,4 +369,12 @@ void Browser::ShowEmojiPanel() {
   ::SendInput(4, input, sizeof(INPUT));
 }
 
+void Browser::ShowAboutPanel() {
+  const auto& opts = about_panel_options_;
+}
+
+void Browser::SetAboutPanelOptions(const base::DictionaryValue& options) {
+  about_panel_options_ = options.Clone();
+}
+
 }  // namespace electron
